change: (Client) Protected Routes & User Auth
Change to ProtectedRoute component, it now waits for Context API to finish fetching the user. Added 'userFound' & 'hasLoaded' states to ProtectedRoute that are used to flag when Context has finished fetching the user.
This is done because 'return outlet' can not be returned from useEffect, and ProtectedRoute is not a syncronus fucntions so using async/await was not possible to await the ContextAPI.