change: (Client) Added Sign In with Magic Link
Added new way to sign user in.
User can now sign in with magic link that will be emailed to them.
Added new input to SignIn page, user can enter an email and request an email with the Sign In link.
User can then click on the link from an email which will hit a auth route and send the link to the Server for validation, if OK user will then be automatically sign in and redirected to the Home page.