added: (Server) Magic LogIn Routes
Added new routes to user:
 - '/magic-link-login'
 - '/login/:token'

that allow user to send a request to create a magic link and send it to an email.
The link, when clicked will send a request to verify the user and if the user is correct sign that user in and send a response to the client.

Added a new email template for:
 - MagicLinkSignIn
 - ForgotPasswordVerification